Initial Commit
Essentially a reincarnation of MMOS, but I fixed a bunch of things, and
I know how to do it ‘right’ this time. The whole compiler (“Make”) was
copied over from MMOS, with a few changes, but I don’t think that I
will have to change it much.
I figured out how to get Eclipse to recognize the c++ stdlib, so a lot
of files are gone. I worked on implementations of things in C++ (as
opposed to assembly), so the code is more readable and is MUCH faster
to write.
Eclipse can now be set up to auto build the C++, and life is good. You
can also open everything in Eclipse at once, so you don’t have to be
running multiple instances to see the whole thing (but you can take
projects offline if you need to conserve memory or something). I mostly
focused on getting stuff to work, mostly the display, instead of
setting up a pretty yet unusable framework that was impossible to
implement. I am going to hold off of porting linux things for a while,
and just work on simple stuff, like low-level display stuff and maybe
the GPIO.