More betterness! Less worseness!
Added entire USPi library, and modified Make to compile it in to the
image. As a side effect of those changes, a lot of bugs regarding
Make’s treatment of static libraries were removed.
A lot of display things were added, including:
 - Character drawing
 - String drawing
 - Integer to stringifying
…Actually that’s it. But it felt like a lot.
It looks pretty good, considering that you can now see letters on the
screen.
In other news, I also made compiling slightly nicer on the mac, by
giving alerts when it works, when it doesn’t work (sometimes), and when
you’ve successfully ejected the SD card (it was annoying me).
I finally finished the manifest system for making static libraries,
where you can specify how to build them with a JSON file.